Make Accordions a single file
